+++ calcipy/code_tag_collector.py
@@ -162,5 +162,5 @@ def write_code_tag_file(
 
     if report:
         path_tag_summary.write_text(f'{header}\n\n{report}\n\n<!-- {SKIP_PHRASE} -->\n')
-    elif path_tag_summary.is_file():
-        path_tag_summary.unlink()
+    else:
+        path_tag_summary.unlink(missing_ok=True)